+++ tests/Deedle.Tests/VirtualParquetSource.fs
@@ -18,6 +18,7 @@ open Deedle.Virtual
 open Deedle.Vectors.Virtual
 open Deedle.Parquet
 open Deedle.Parquet.Virtual.Sources
+open Deedle.TestData
 open Deedle.Virtual.Sources
 open Deedle.Tests.VirtualInstrumentation